feat: Add exo_playback_control_view.xml to override layout files for ExoPlayer
Overriding layout files
When these views are inflated, they use specific layout files to determine
what the UI looks like. For SimpleExoPlayerView, this file is called:
exo_simple_player_view.xml. This layout file includes a PlayBackControlView
(once it's inflated, it replaces the exo_controller_placeholder item) which
also uses its own layout file: exo_playback_control_view.xml.

If you include any layout files with the same names, ExoPlayer will use them
instead of these default ones. This allows you to fully customize what the UI
looks like.

One caveat: Use of standard ids (such as exo_play) is required so that child
views can be identified, bound to the player and updated in an appropriate
way. A full list of the standard ids for each view can be found in the Javadoc
for PlaybackControlView and SimpleExoPlayerView. Use of each standard id is
optional, which is why we’ll be able to omit some of the usual playback
controls in our example.